Made url a required prop and combined Chiclet test files
