+++ packages/experiments/src/components/Chiclet/ChicletXsmall.base.tsx
@@ -29,7 +29,7 @@ export class ChicletXsmallBase extends React.Component<IChicletCardProps, {}> {
         {generatePreview(this.props, imageProvided, this._classNames, PREVIEW_IMAGE_HEIGHT, PREVIEW_IMAGE_WIDTH)}
         <div className={this._classNames.info}>
           <div className={this._classNames.title}>{title ? title : null}</div>
-          <div className={this._classNames.url}>{url ? url : null}</div>
+          <div className={this._classNames.url}>{url}</div>
         </div>
         {footer}
       </div>
